+++ tests/test_binnedloss.py
@@ -158,7 +158,7 @@ def test_binned_loss(weights, Loss, simultaneous):
     test_values = znp.array(test_values_np_shifted)
     test_values = zfit.Data.from_tensor(obs=obs, tensor=test_values, weights=weights)
     init_yield = test_values_np.shape[0] * 1.2
-    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, step_size=1)
+    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, stepsize=1)
     binning = zfit.binned.RegularBinning(92, obs.v1.lower, obs.v1.upper, name="obs1")
     obs_binned = obs.with_binning(binning)
     test_values_binned = test_values.to_binned(obs_binned)
@@ -288,7 +288,7 @@ def test_binned_chi2_loss(Loss, empty, errors):  # TODO: add test with zeros in 
     test_values = znp.array(test_values_np_shifted)
     test_values = zfit.Data.from_tensor(obs=obs, tensor=test_values)
     init_yield = test_values_np.shape[0] * 1.2
-    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, step_size=1)
+    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, stepsize=1)
     binning = zfit.binned.RegularBinning(32, obs.v1.lower, obs.v1.upper, name="obs1")
     obs_binned = obs.with_binning(binning)
     test_values_binned = test_values.to_binned(obs_binned)
@@ -326,7 +326,7 @@ def test_binned_loss_hist(weights, Loss):
     test_values = znp.array(test_values_np_shifted)
     test_values = zfit.Data.from_tensor(obs=obs, tensor=test_values, weights=weights)
     init_yield = test_values_np.shape[0] * 1.2
-    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, step_size=1)
+    scale = zfit.Parameter("yield", init_yield, 0, init_yield * 4, stepsize=1)
     binning = zfit.binned.RegularBinning(32, obs.v1.lower, obs.v1.upper, name="obs1")
     obs_binned = obs.with_binning(binning)
     test_values_binned = test_values.to_binned(obs_binned)