+++ zfit/minimizers/minimizer_minuit.py
@@ -285,18 +285,18 @@ class Minuit(BaseMinimizer, GraphCachable):
             name=params_name,
         )
         minimizer.precision = precision
-        approx_step_sizes = {}
+        approx_stepsizes = {}
         # get possible initial step size from previous minimizer
         if init:
-            approx_step_sizes = init.hesse(params=params, method="approx", name="approx")
+            approx_stepsizes = init.hesse(params=params, method="approx", name="approx")
 
         empty_dict = {}
         for param in params:
-            step_size = approx_step_sizes.get(param, empty_dict).get("error")
-            if step_size is None and param.has_step_size:
-                step_size = param.stepsize
-            if step_size is not None:
-                minimizer.errors[param.name] = step_size
+            stepsize = approx_stepsizes.get(param, empty_dict).get("error")
+            if stepsize is None and param.has_stepsize:
+                stepsize = param.stepsize
+            if stepsize is not None:
+                minimizer.errors[param.name] = stepsize
         # set limits
         for param in params:
             if param.has_limits: